Moved storeConfigs to zustand state
